% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2IsoMErData.R
\docType{data}
\name{exact_masses}
\alias{exact_masses}
\title{Descriptive data title}
\format{
a \code{data.table} containing the following fields:
\describe{
\item{formula}{}
\item{posH}{}
\item{posNa}{}
\item{posK}{}
\item{negH}{}
\item{negCl}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
