% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2IsoMErData.R
\docType{data}
\name{example_MSEA_multicond}
\alias{example_MSEA_multicond}
\title{Descriptive data title}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{Term}{}
\item{n}{}
\item{ES_median}{}
\item{ES_sd}{}
\item{p.value_combined}{}
\item{q.value_combined}{}
\item{fraction.bootstrap.presence}{}
\item{condition.x}{}
\item{condition.y}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
