% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2IsoMErData.R
\docType{data}
\name{example_MSEA_obj}
\alias{example_MSEA_obj}
\title{Descriptive data title}
\format{
a \code{S2IsoMEr} containing the following fields:
\describe{
\item{scmatrix}{}
\item{polarization_mode}{}
\item{enrichment_type}{}
\item{Annotation_database}{}
\item{Custom_database}{}
\item{annotations}{}
\item{annotation.weights}{}
\item{isobars_list}{}
\item{conditions}{}
\item{include}{}
\item{consider_isomers}{}
\item{consider_isobars}{}
\item{mass_range_ppm}{}
\item{background_name}{}
\item{endogenous_only}{}
\item{pathway_assoc_only}{}
\item{remove_expected_predicted}{}
\item{pathway_list}{}
\item{LUT}{}
\item{termsSelection}{}
\item{condition.x}{}
\item{condition.y}{}
\item{ranking.by}{}
\item{gsea.method}{}
\item{rankings}{}
\item{enrichment_analysis}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
