% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2IsoMErData.R
\docType{data}
\name{example_ORA_custom_universe}
\alias{example_ORA_custom_universe}
\title{Descriptive data title}
\format{
a \code{character} containing the following fields:
\describe{
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
