% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2IsoMErData.R
\docType{data}
\name{example_ORA_markers}
\alias{example_ORA_markers}
\title{Descriptive data title}
\format{
a \code{list} containing the following fields:
\describe{
\item{upregulated}{}
\item{downregulated}{}
\item{all}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
