% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2IsoMErData.R
\docType{data}
\name{LION_LUT}
\alias{LION_LUT}
\title{Descriptive data title}
\format{
a \code{data.table} containing the following fields:
\describe{
\item{ID}{}
\item{name}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
