% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2IsoMErData.R
\docType{data}
\name{LipidMaps_pathway_name}
\alias{LipidMaps_pathway_name}
\title{Descriptive data title}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
