% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2IsoMErData.R
\docType{data}
\name{Metabo_class_sf}
\alias{Metabo_class_sf}
\title{Descriptive data title}
\format{
a \code{list} containing the following fields:
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
