% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2IsoMErData.R
\docType{data}
\name{metaspace_databases}
\alias{metaspace_databases}
\title{Descriptive data title}
\format{
a \code{data.table} containing the following fields:
\describe{
\item{formula}{}
\item{name}{}
\item{db}{}
\item{Endogenous}{}
\item{Pathway_assoc}{}
\item{HMDB_status}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
