% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2IsoMErData.R
\docType{package}
\name{S2IsoMErData}
\alias{S2IsoMErData}
\alias{S2IsoMErData-package}
\title{Package Title}
\description{
A description of the data package
}
\details{
S2IsoMErData
A data package for S2IsoMErData.

Use \code{data(package='S2IsoMErData')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\author{
\strong{Maintainer}: Bishoy Wadie \email{bishoy.wadie@embl.de} (\href{https://orcid.org/0000-0001-8692-1107}{ORCID})

Authors:
\itemize{
  \item Martijn Molenaar \email{m.r.molenaar@protonmail.com} (\href{https://orcid.org/0000-0001-5221-608X}{ORCID})
}

Other contributors:
\itemize{
  \item Theodore Alexandrov \email{talexandrov@health.ucsd.edu} (\href{https://orcid.org/0000-0001-9464-6125}{ORCID}) [copyright holder, funder]
}

}
