% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2IsoMErData.R
\docType{data}
\name{termsSelection}
\alias{termsSelection}
\title{Descriptive data title}
\format{
a \code{data.frame} containing the following fields:
\describe{
\item{LION_ID}{}
\item{LION_name}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
